requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
Django==3.2.3
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-guardian==2.4.0
django-phonenumber-field==5.2.0
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.9
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
drf-yasg==1.20.0
future==0.18.2
gunicorn==20.1.0
huepy==1.2.1
idna==2.10
imageio==2.13.2
imageio-ffmpeg==0.4.5
inflection==0.5.1
iniconfig==1.1.1
instabot==0.117.0
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
mock==4.0.3
moviepy==1.0.3
numpy==1.21.4
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==20.9
phonenumbers==8.12.24
Pillow==8.4.0
pluggy==1.0.0
proglog==0.1.9
psycopg2-binary==2.8.6
py==1.11.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.5
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.16.0
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
schedule==1.1.0
simplejson==3.17.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
tqdm==4.62.3
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.5
